separate PartitionWriter into interface and Disk implementation
