+++ .java-version
@@ -1,0 +1,1 @@
+11.0
