+++ src/main/java/de/azapps/kafkabackup/common/partition/PartitionException.java
@@ -1,0 +1,10 @@
+package de.azapps.kafkabackup.common.partition;
+
+
+public class PartitionException extends Exception {
+    public PartitionException(String message) {
+        super(message);
+    }
+    public PartitionException(Throwable e) { super(e); }
+    public PartitionException(String message, Throwable e) { super(message, e); }
+}
