+++ src/main/java/de/azapps/kafkabackup/common/partition/disk/PartitionIndexRestore.java
@@ -1,4 +1,4 @@
-package de.azapps.kafkabackup.common.partition;
+package de.azapps.kafkabackup.common.partition.disk;
 
 import de.azapps.kafkabackup.common.segment.SegmentUtils;
